# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: api-lifecare-integrator
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "1.0"
servers:
- url: http://localhost:61256
  description: Generated server url
...
paths:
...
  /{municipalityId}/users:
    get:
      tags:
      - Users
      description: Get the family care caseworkers (users)
      operationId: getUsers
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: limit
        in: query
        description: The maximum number of users to return
        required: true
        schema:
          type: integer
          format: int32
          exclusiveMinimum: 0
          maximum: 1000
        example: 100
      - name: offset
        in: query
        description: The starting point within the result set
        required: false
        schema:
          type: integer
          format: int32
        example: 0
      - name: modifiedAfter
        in: query
        description: Only users modified after this time
        required: false
        schema:
          type: string
          format: date-time
      - name: modifiedBefore
        in: query
        description: Only users modified before this time
        required: false
        schema:
          type: string
          format: date-time
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseworkerUser"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/person:
    get:
      tags:
      - Person
      description: Get person master data for a person.
      operationId: getPerson
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: partyId
        in: query
        description: Party id of the person
        required: true
        schema:
          type: string
        example: 81471222-5798-11e9-ae24-57fa13b361e1
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Person"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/documents:
    get:
      tags:
      - Documents
      description: Get document metadata for a person. The content of a single document
        is fetched separately
      operationId: getDocuments
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: partyId
        in: query
        description: Party id of the person
        required: true
        schema:
          type: string
        example: 81471222-5798-11e9-ae24-57fa13b361e1
      - name: from
        in: query
        description: Only include documents dated on or after this date
        required: false
        schema:
          type: string
          format: date
        example: 2025-01-01
      - name: to
        in: query
        description: Only include documents dated on or before this date
        required: false
        schema:
          type: string
          format: date
        example: 2026-12-31
      - name: page
        in: query
        description: The page to fetch (one-based)
        required: false
        schema:
          type: integer
          format: int32
          exclusiveMinimum: 0
        example: 1
      - name: pageSize
        in: query
        description: The number of records per page
        required: false
        schema:
          type: integer
          format: int32
          exclusiveMinimum: 0
          maximum: 1000
        example: 20
      - name: ascending
        in: query
        description: Whether to sort ascending
        required: false
        schema:
          type: boolean
        example: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResponseDocumentMetadata"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/documents/{documentId}/content:
    get:
      tags:
      - Documents
      description: Get the content of a document as a PDF
      operationId: getDocumentContent
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: documentId
        in: path
        description: The document id
        required: true
        schema:
          type: string
          minLength: 1
        example: 12345
      responses:
        "200":
          description: Successful operation
          content:
            application/pdf:
              schema:
                type: string
                format: byte
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/contacts:
    get:
      tags:
      - Person
      description: Get the contacts attached to a person.
      operationId: getContacts
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: partyId
        in: query
        description: Party id of the person
        required: true
        schema:
          type: string
        example: 81471222-5798-11e9-ae24-57fa13b361e1
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
...
    ConstraintViolationProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        causeAsProblem:
          $ref: "#/components/schemas/ThrowableProblem"
        detail:
          type: string
        instance:
          type: string
          format: uri
    ThrowableProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        causeAsProblem: {}
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    CaseworkerUser:
      type: object
      description: A caseworker user account in the Lifecare family care system
      properties:
        id:
          type: string
          description: The user id in the source system
          examples:
          - "12345"
        hsaId:
          type: string
          description: The HSA id of the user
          examples:
          - SE2321000032-1234
        networkUserId:
          type: string
          description: The network user id of the user
          examples:
          - annand01
        firstName:
          type: string
          description: The first name of the user
          examples:
          - Anna
        lastName:
          type: string
          description: The last name of the user
          examples:
          - Andersson
        fullName:
          type: string
          description: The full name of the user
          examples:
          - Anna Andersson
        description:
          type: string
          description: A description of the user
          examples:
          - Handläggare ekonomiskt bistånd
        validFrom:
          type: string
          format: date-time
          description: The date and time the user account is valid from
          examples:
          - 2026-01-01T00:00:00Z
        validTo:
          type: string
          format: date-time
          description: The date and time the user account is valid to
          examples:
          - 2026-12-31T23:59:59Z
        disabled:
          type: boolean
          description: Whether the user account is disabled
          examples:
          - "false"
...
    Person:
      type: object
      description: Person master data from Lifecare family care
      properties:
        customerNumber:
          type: integer
          format: int32
          description: The customer number in the source system
          examples:
          - 12345
        name:
          type: string
          description: Full name of the person
          examples:
          - Anna Andersson
        streetAddress:
          type: string
          description: Street address
          examples:
          - Storgatan 1
        careOfAddress:
          type: string
          description: Care of (c/o) address
          examples:
          - c/o Karlsson
        postalCode:
          type: string
          description: Postal code
          examples:
          - "85185"
        postalAddress:
          type: string
          description: Postal address
          examples:
          - Sundsvall
        phoneHome:
          type: string
          description: Home phone number
          examples:
          - 060-123456
        phoneWork:
          type: string
          description: Work phone number
          examples:
          - 060-654321
        phoneMobile:
          type: string
          description: Mobile phone number
          examples:
          - 070-1234567
        email:
          type: string
          description: Email address
          examples:
          - anna.andersson@example.com
        addressProtection:
          type: boolean
          description: Whether the person's address is protected
          examples:
          - "false"
        secretPhone:
          type: boolean
          description: Whether the person's phone number is secret
          examples:
          - "false"
        protectedRegistration:
          type: boolean
          description: Whether the person has protected registration
          examples:
          - "false"
...
    DocumentMetadata:
      type: object
      description: Metadata for a document stored in the Lifecare family care system.
        The document content is fetched separately
      properties:
        id:
          type: string
          description: The document id in the source system
          examples:
          - "12345"
        title:
          type: string
          description: The document title
          examples:
          - Utredning ekonomiskt bistånd
        date:
          type: string
          format: date
          description: The document date
          examples:
          - 2026-05-01
        documentType:
          type: string
          description: The document type
          examples:
          - Utredning
        ownerType:
          type: string
          description: The type of the document owner
          examples:
          - Person
    PagedResponseDocumentMetadata:
      type: object
      description: "Paged response wrapper, mirroring the pagination of the underlying\
        \ Lifecare family care API"
      properties:
        page:
          type: integer
          format: int32
          description: The current page (one-based)
          examples:
          - 1
        pageSize:
          type: integer
          format: int32
          description: The number of records per page
          examples:
          - 20
        totalPages:
          type: integer
          format: int32
          description: The total number of pages
          examples:
          - 5
        totalRecords:
          type: integer
          format: int32
          description: The total number of records
          examples:
          - 98
        results:
          type: array
          description: The records on the current page
          items:
            $ref: "#/components/schemas/DocumentMetadata"
...
    Address:
      type: object
      description: Address of a contact
      properties:
        visitingAddress:
          type: string
          description: Visiting address
          examples:
          - Storgatan 1
        streetAddress:
          type: string
          description: Street address
          examples:
          - Storgatan 1
        postalCode:
          type: string
          description: Postal code
          examples:
          - "85185"
        postalAddress:
          type: string
          description: Postal address
          examples:
          - Sundsvall
    Contact:
      type: object
      description: A contact (e.g. caseworker) attached to a person in Lifecare family
        care
      properties:
        id:
          type: string
          description: Id of the contact
          examples:
          - "12345"
        name:
          type: string
          description: Name of the contact
          examples:
          - Bo Bosson
        employmentTitle:
          type: string
          description: Employment title of the contact
          examples:
          - Socialsekreterare
        organizationName:
          type: string
          description: Organization name of the contact
          examples:
          - Ekonomiskt bistånd
        phone:
          type: string
          description: Phone number
          examples:
          - 060-123456
        email:
          type: string
          description: Email address
          examples:
          - bo.bosson@example.com
        address:
          $ref: "#/components/schemas/Address"
          description: Address of the contact
        typeOfContact:
          type: string
          description: Type of contact
          examples:
          - Handläggare
...
  securitySchemes: {}